Finalize CPP09: clean up ex02, add study guide PDF
- ex02: restructure FordJohnson into .hpp/.cpp split, remove debug
  code and test artifacts, add ALGORITHM.md documentation
- ex00/ex01: minor cleanup
- Add CPP09_study.pdf covering all three exercises

+++ ex01/RPN.hpp
@@ -1,13 +1,14 @@
 #pragma once
 
 #include <iostream>
+#include <sstream>
 #include <stack>
 #include <string>
 
 class RPN {
     private:
         std::string _input;
-        std::stack<int> _stack;
+        std::stack<long long> _stack;
     public:
         RPN();
         RPN(const std::string& _inputval);
